src/client/requirements.txt
mss==9.0.1
numpy==1.26.4
pillow==10.2.0
scipy==1.12.0